# cache taps