--- a/Library/Homebrew/cmd/tap-pin.rb
+++ b/Library/Homebrew/cmd/tap-pin.rb
@@ -3,7 +3,7 @@ require "cmd/tap"
 module Homebrew
   def tap_pin
     ARGV.named.each do |name|
-      tap = Tap.new(*tap_args(name))
+      tap = Tap.fetch(*tap_args(name))
       tap.pin
       ohai "Pinned #{tap.name}"
     end