--- a/Library/Homebrew/cmd/tap-unpin.rb
+++ b/Library/Homebrew/cmd/tap-unpin.rb
@@ -3,7 +3,7 @@ require "cmd/tap"
 module Homebrew
   def tap_unpin
     ARGV.named.each do |name|
-      tap = Tap.new(*tap_args(name))
+      tap = Tap.fetch(*tap_args(name))
       tap.unpin
       ohai "Unpinned #{tap.name}"
     end